+++ Cron/CancelOrders.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 namespace Aplazo\AplazoPayment\Cron;
 
 use Aplazo\AplazoPayment\Model\Service\OrderService;
+use Aplazo\AplazoPayment\Service\ApiService;
 use Magento\Sales\Model\Order;
 
 class CancelOrders
@@ -23,16 +24,23 @@ class CancelOrders
     private $orderService;
 
     /**
-     * Constructor
-     *
-     * @param \Psr\Log\LoggerInterface $logger
+     * @var ApiService
+     */
+    private $apiService;
+
+    /**
+     * @param OrderService $orderService
+     * @param ApiService $apiService
+     * @param \Aplazo\AplazoPayment\Helper\Data $aplazoHelper
      */
     public function __construct(
         OrderService $orderService,
+        ApiService $apiService,
         \Aplazo\AplazoPayment\Helper\Data $aplazoHelper
     )
     {
         $this->orderService = $orderService;
+        $this->apiService   = $apiService;
         $this->aplazoHelper = $aplazoHelper;
     }
 
@@ -57,15 +65,36 @@ class CancelOrders
                     $this->aplazoHelper->log('Total de ordenes encontradas: ' . $orderCollectionCount);
 
                     $incrementId = $order->getIncrementId();
-                    $cancelResponse = $this->orderService->cancelOrder($order->getId());
-                    if($cancelResponse['success']){
-                        $this->aplazoHelper->log("Aplazo Cancel Orders: StoreId $store_id - $counter/$orderCollectionCount - #$incrementId - Cancelada exitosamente");
-                        $ordersCanceledCount++;
+                    if($this->apiService->getLoanStatus($incrementId)){
+                        try {
+                            $orderResult = $this->orderService->getOrderByIncrementId($incrementId);
+                            if ($orderResult['success']) {
+                                /** @var Order $order */
+                                $order = $orderResult['order'];
+                                $order = $this->orderService->approveOrder($order->getId());
+                                $orderPayment = $order->getPayment();
+                                $orderPayment->setAdditionalInformation('aplazo_status', $this->apiService::LOAN_SUCCESS_STATUS);
+                                $order->addCommentToStatusHistory('Orden en Aplazo pagada correctamente. Notificación realizada a través de cron.');
+                                $this->orderService->saveOrder($order);
+                                $this->aplazoHelper->log("Order $incrementId is OUTSTANDING in Aplazo.");
+                            } else {
+                                $this->aplazoHelper->log('Order incrementId not found ' . $incrementId);
+                            }
+                        } catch (\Exception $e) {
+                            $this->aplazoHelper->log('Order could not advance to paid status: ' . $incrementId) .'. '. $e->getMessage();
+                        }
+                    } else {
+                        $cancelResponse = $this->orderService->cancelOrder($order->getId());
+                        if($cancelResponse['success']){
+                            $this->aplazoHelper->log("Aplazo Cancel Orders: StoreId $store_id - $counter/$orderCollectionCount - #$incrementId - Cancelada exitosamente");
+                            $ordersCanceledCount++;
+                        }
+                        else{
+                            $this->aplazoHelper->log("<comment>Aplazo Cancel Orders: StoreId $store_id - $counter/$orderCollectionCount - #$incrementId - Con detalles</comment>");
+                            $ordersWithErrors[] = $incrementId . ' ' . $cancelResponse['message'];
+                        }
                     }
-                    else{
-                        $this->aplazoHelper->log("<comment>Aplazo Cancel Orders: StoreId $store_id - $counter/$orderCollectionCount - #$incrementId - Con detalles</comment>");
-                        $ordersWithErrors[] = $incrementId . ' ' . $cancelResponse['message'];
-                    }
+
                     $counter++;
                 }
             }