+++ versionhistory.md
@@ -110,3 +110,9 @@
 * Object manager removed in product and cart page.
 ## 3.2.4
 * Aplazo cartId changed by magento increment_id.
+## 3.2.5
+* Check order status in Aplazo before cancel in cron. Logs added in webhook.
+## 3.2.6
+* Not MSI stock.
+## 3.2.7
+* Success controller fix failure redirect.